requirements.txt
MetaTrader5
pandas
numpy
joblib
tensorflow
pandas_ta
streamlit
plotly
feedparser
requests
python-dotenv
yfinance
